Developer (#1)
* chore: Update project dependencies in pyproject.toml

* feat: Add MongoDB connection settings and tests

* chore: Update pymongo dependency to version 4.8.0 in pyproject.toml

* chore: Update MongoDB image in docker-compose.yml

* chore: Refactor MongoDB connection settings and tests

* feat: Add OrderRepositoryInterface and OrderRepository

This commit adds the `OrderRepositoryInterface` and `OrderRepository` classes to the codebase. The `OrderRepositoryInterface` defines the interface for interacting with order data, while the `OrderRepository` implements the interface using MongoDB as the underlying database. These changes provide the foundation for managing orders in the system.

* feat: Add Order and Product models

This commit adds the `Order` and `Product` models to the codebase. The `Order` model represents a purchase order and includes fields such as `order_id`, `client_id`, `product_list`, `total`, `created_at`, and `updated_at`. The `Product` model represents a product and includes fields such as `product_id`, `name`, `price`, `created_at`, and `updated_at`. These models provide the foundation for managing orders and products in the system.

* feat: Rename get_order method to get_order_by_id in OrderRepositoryInterface

This commit updates the `OrderRepositoryInterface` by renaming the `get_order` method to `get_order_by_id`. This change provides a more descriptive and accurate method name, improving the clarity and readability of the codebase.

* feat: Add purchase_order_microservice/app.py

This commit adds the `purchase_order_microservice/app.py` file to the codebase. This file contains the initial implementation of the purchase order microservice.

* feat: Add references to README.md

This commit adds references to the README.md file. The references provide links to external resources related to MongoDB, FastAPI, and Pydantic. This addition enhances the documentation and provides additional information for developers working with the codebase.

* feat: Skip sensitive tests in order repository and connection settings